migrating new color component to all the right tab panels
