+++ packages/component-style-layer/src/components/commons/FlexChildInput.tsx
@@ -1,4 +1,3 @@
-import { gray100, gray800, smallText } from "@atrilabs/design-system";
 import React from "react";
 import { CssProprtyComponentType } from "../../types";
 import ControlledInput from "./ControlledInput";
@@ -8,20 +7,6 @@ export type FlexChildInputProps = {
   patchCb: CssProprtyComponentType["patchCb"];
   styles: CssProprtyComponentType["styles"];
   defaultValue: string | number;
-};
-
-const styles: { [key: string]: React.CSSProperties } = {
-  inputBox: {
-    ...smallText,
-    outline: "none",
-    color: gray100,
-    backgroundColor: gray800,
-    width: "30px",
-    border: "none",
-    borderRadius: "2px",
-    textAlign: "center",
-    lineHeight: "25px",
-  },
 };
 
 export const FlexChildInput: React.FC<FlexChildInputProps> = (props) => {
@@ -39,17 +24,6 @@ export const FlexChildInput: React.FC<FlexChildInputProps> = (props) => {
   };
   return (
     <div>
-      {/* <input
-        type="text"
-        value={props.styles[props.styleItem] || ""}
-        onChange={(e) => handleChange(e, props.styleItem)}
-        style={styles.inputBox}
-        placeholder={
-          typeof props.defaultValue === "number"
-            ? props.defaultValue.toString()
-            : props.defaultValue
-        }
-      /> */}
       <ControlledInput
         type="text"
         value={props.styles[props.styleItem] || ""}
